Fix XSSFExcelExtractor to emit empty cells
The default IRow enumerator will skip empty cells. This makes it so
that the output from the Text property has cells that don't line up
with their corresponding headers.

This fix calls GetCell instead, which will return null in cases where
the cell was empty. The tab delimiters work as expected now.