Fix thread safety issue in sanity checker test
This was causing issues when testing on unix.
Check was not being assigned a value, and resulted
in a null reference exception when run. This change
fixes the missing check data, and removes the use
of Threads, as data is not passed in a safe manner
between the calling thread and the created thread.