Get a copy of the CT_Border when about to mutate it
GetCTBorder() returns a shared instance when applyBorder is true. When GetCTBorder() is invoked from inside a property setter, that shared instance is then mutated and all other XSSFCellStyle instances that referred to the original (before mutation) CT_Border are changed as well. This appears to be unintentional. If we know we are inside a setter (and therefore about to mutate the state of the CT_Border) then we should perform that mutation on a copy.