Re-worked custom XSSFCellUnit
There were a lot of problems with it, it was very hard to quantify the changes into different commits, so sorry for the confusion. This class needs to be rebuilt from the gorund up. Or better yet rid of completely. Main problem that this class had, that wass fixed - it was writing properties to a style one way, and reading them another way. Thus every lookup of the styles was failing, causing thousands of duplicate styles and ungodly slow performances.