Fixes bugs revealed by testing
Got rid of default character height, it as needed by width getter, because excel limits width of the column by number of characters, and height isn't tied to it. Fixed how row span is counted, it was wrong. Added ratio to final height so it is close to how  excel would count it.