Upstream-fix: Add AddColor method to CT_Font
XSSFFontFormatting.FontColor setter used to set a new color to the CT_Font/colorField list which was not always initialised. Thsi adds a method which first will check if the list is initialized and then adds a color to it. XSSFFontFormatting.FontColor setter now uses this method to set a color.