Refactor XSSFSheet.ShiftedRowNum private method
Rename to XSSFSheet.ShiftedRowOrColumnNumber, rename parameters to not be tied to only rows.

Refactor XSSFSheet.ShouldRemoveRow private method

Rename to XSSFSheet.ShouldRemoveAtIndex, rename parapmeters to not be tied to rows.

Rename XSSFSheet.RemoveOverwritten private method to XSSFSheet.RemoveOverwrittenRows

Refactor XSSFSheet.HeadMapCount method to be more general.
To be later able to use it fot columns.