Fixing a problem when the shared string table record reports more strings than are present in the file.
The current behaviour is: read the number of strings in the string table. Then attempt to read that many number of strings. Using subsequent records if needed. If there are less strings present than reported, pad the internal string table with empty strings. This only works when there are no more records in the stream.

This fix adds a check to ensure that the next record is a continuation record. If it is not then the internal string table is padded as before.